fix: fix rendering issue in Firefox
This fix moves the responsibility of the rendering into the extension, rather than the mule-preview
plugin. An issue was uncovered  where the latest version of React (16.9.0) does not work in Firefox
browser extensions due to a change to the scheduler (https://github.com/facebook/react/issues/16606)
so React was downgraded. After those fixes were implemented, the issue with rendering only working
correctly the second time the extension was triggered was fixed.

fix #72